% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqcheck.R
\name{seqcheck}
\alias{seqcheck}
\title{DNA/RNA sequence checker}
\usage{
seqcheck(seq)
}
\arguments{
\item{seq}{Input sequence (string)}
}
\description{
This function checks whether the input sequence is a DNA sequence or a RNA sequence.
}
\details{
If the input sequence includes any letters which are not "A", "T", "U", "C", or "G", it will be recognised as invalid.
}
\examples{
seqcheck("ATCG") #A DNA sequence
seqcheck("AUCG") #A RNA sequence
seqcheck("AJEI") #An invalid sequence
seqcheck("atcg") #the function works with both cases
}
\keyword{DNA,}
\keyword{RNA,}
\keyword{sequence}
