% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqcomplement.R
\name{seqcomplement}
\alias{seqcomplement}
\title{Complementary/Reverse complementary sequence generator}
\usage{
seqcomplement(seq, uppercase = TRUE, reverse = FALSE)
}
\arguments{
\item{seq}{Input sequence (string)}

\item{uppercase}{TRUE (default) or FALSE, determining whether the output sequence is printed in uppercase (TRUE) or lowercase (FALSE) letters.}

\item{reverse}{TRUE or FALSE (default), determining whether the output sequence is the complemetary sequence(FALSE) or the reverse complementary sequence (TRUE).}
}
\description{
This function converts an input DNA or RNA sequence into it's complementary or reverse complementary sequence.
}
\details{
If the input sequence includes any letters which are not "A", "T", "U", "C", or "G", it will be recognised as invalid.
}
\examples{
seqcomplement("ATCG", reverse = TRUE) #A reverse complementary DNA sequence 
seqcomplement("AUCG", uppercase = FALSE) #A lowercase complementary RNA sequence
seqcomplement("AJEI") #An invalid sequence
}
\keyword{DNA,}
\keyword{RNA,}
\keyword{complementary}
\keyword{reverse}
\keyword{sequence}
\keyword{sequence,}
