% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqcomposition.R
\name{seqcomposition}
\alias{seqcomposition}
\title{Sequence Composition Finder}
\usage{
seqcomposition(seq, type = "ALL")
}
\arguments{
\item{seq}{Input sequence (string)}

\item{type}{"GC", "A", "T", "C", "G", "U", "ALL" (default). The type of composition that should be returned. "GC" returns the GC content. "A" returns the percentage of adenine bases present ("T", "G", "C", "U" returns the respective base percentages). "ALL" returns all percentages and the GC content.}
}
\description{
This function returns the GC-content or the percentage of a certain base in a given sequence.
}
\details{
If the input sequence includes any letters which are not "A", "T", "U", "C", or "G", it will be recognised as invalid.
}
\examples{
seqcomposition("ATCG", "A") #percentage adenine bases
seqcomposition("ATCG", "T") #percentage adenine bases
seqcomposition("ATCG", "C") #percentage adenine bases
seqcomposition("ATCG", "G") #percentage adenine bases
seqcomposition("AUCG", "U") #percentage adenine bases
seqcomposition("ATCG", "ALL") #GC-content and percentages of bases
}
\keyword{composition,}
\keyword{gc-content,}
\keyword{percentage}
\keyword{sequence,}
