% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqlength.R
\name{seqlength}
\alias{seqlength}
\title{Sequence Length Checker}
\usage{
seqlength(seq)
}
\arguments{
\item{seq}{Input sequence (string)}
}
\description{
This function checks the length of an input sequence
}
\details{
This function can check the length of sequences (as strings).
}
\examples{
seqlength("ATCG") #Sequence must be a string
}
\keyword{length}
\keyword{sequence,}
