% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqlocator.R
\name{seqlocator}
\alias{seqlocator}
\title{Sequence locator}
\usage{
seqlocator(main, target)
}
\arguments{
\item{main}{Input sequence (string)}

\item{target}{Target sequence (string)}
}
\description{
This function returns the number of instances of a specified sequence within an input sequence and where they occur.
}
\details{
If the input sequence includes any letters which are not "A", "T", "U", "C", or "G", it will be recognised as invalid.
}
\examples{
seqlocator("ATCGATCG", "AT") #counts number of "AT" in the "ATCGATCG" sequence
seqlocator("ATCGATCG", "AC") #no instances of "AC" in the sequence
}
\keyword{instances,}
\keyword{positions}
\keyword{sequence,}
\keyword{target_sequence,}
