% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqsummary.R
\name{seqsummary}
\alias{seqsummary}
\title{Sequence Summary}
\usage{
seqsummary(seq)
}
\arguments{
\item{seq}{Input sequence (string)}
}
\description{
This function returns a summary of a given sequence, including the type of sequence, the length of the sequence, and the percentages of bases present.
}
\details{
If the input sequence includes any letters which are not "A", "T", "U", "C", or "G", it will be recognised as invalid.
}
\examples{
seqsummary("ATCG") #Returns a summary of the sequence
}
\keyword{DNA,}
\keyword{RNA,}
\keyword{composition,}
\keyword{gc-content,}
\keyword{length,}
\keyword{percentage}
\keyword{sequence,}
