% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqtranscribe.R
\name{seqtranscribe}
\alias{seqtranscribe}
\title{Sequence transcription}
\usage{
seqtranscribe(seq, type)
}
\arguments{
\item{seq}{Input sequence (string)}

\item{type}{"DNA", "RNA". The target sequence type to be transcribed to.}
}
\description{
This function transcribes a given sequence into RNA or DNA
}
\examples{
seqtranscribe("AUCG", type = "DNA") #transcribes a RNA sequence into DNA
seqtranscribe("ATCG", type = "RNA") #transcribes a DNA sequence into RNA
seqtranscribe("ASTQ", type = "DNA") #Invalid Sequence
}
\keyword{DNA,}
\keyword{RNA}
\keyword{sequence,}
\keyword{transcribe,}
