% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqtranslate.R
\name{seqtranslate}
\alias{seqtranslate}
\title{Sequence translator}
\usage{
seqtranslate(seq, type = 1)
}
\arguments{
\item{seq}{Input sequence (string)}

\item{type}{"1", "3". Output sequence type: 1-letter (1)(Default) or 3-letter (3)}
}
\description{
This function translates an RNA sequence into an amino acid sequence (expressed in one letter or three letter code).
}
\details{
The function does not recognise DNA sequences. If the input sequence includes any letters which are not "A", "T", "U", "C", or "G", it will be recognised as invalid.
}
\examples{
seqtranslate("AAAUUUCCCGGG", type = 1) #Translating an RNA sequence into 1-letter amino acid code
seqtranslate("AAAUUUCCCGGG", type = 3) #Translating an RNA sequence into 3-letter amino acid code
seqtranslate("AAATTTCCCGGG", type = 1) #DNA sequences are invalid
}
\keyword{1-letter}
\keyword{3-letter}
\keyword{RNA}
\keyword{acid}
\keyword{amino}
\keyword{code}
\keyword{code,}
\keyword{primary}
\keyword{sequence,}
\keyword{structure,}
